requirements.txt
python-telegram-bot==20.5
python-dotenv
langchain
pytesseract
Pillow
gigachat
